ingestion/db_client.py
"""
Shared SQLAlchemy engine and session factory.
Imported by ingestion/, agent_tools/, and api/ modules.
"""
import os
from sqlalchemy import create_engine
from sqlalchemy.orm import sessionmaker
from dotenv import load_dotenv

load_dotenv()

DATABASE_URL = os.environ.get("DATABASE_URL", "postgresql://pocketia:pocketia@localhost:5432/pocketia")

engine = create_engine(DATABASE_URL, pool_pre_ping=True, pool_size=5, max_overflow=10)
SessionLocal = sessionmaker(bind=engine, autoflush=False, autocommit=False)
