requirements.txt
# Core AI
anthropic>=0.40.0

# Slack
slack-bolt>=1.18.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Data ingestion
pandas>=2.2.0

# Scheduling
apscheduler>=3.10.0

# Config
python-dotenv>=1.0.0
